test(agent): izolează testele de buclă de stratul real de transport
test(agent): izolează testele de buclă de stratul real de transport

Testele de buclă înlocuiau doar send_heartbeat, lăsând check_server_health și
register_agent pe implementările reale. Ele sunt accesibile din heartbeat_loop
fără a fi numite direct: directiva reregister și ramura AgentNotRegisteredError
trec amândouă prin register_agent_with_retry.

Prima modificare a unui răspuns fals ar fi scos testul unitar în rețea. Bucla de
startup nu abandonează niciodată din cauza numărului de încercări, iar
stop_event se setează doar din send_heartbeat, care nu mai e apelat pe acea
cale, deci suita ar fi atârnat definitiv contra unui port închis. Cu serverul de
dezvoltare pornit local, ar fi făcut în schimb o înregistrare reală, cu scriere
în baza lui de date - rezultat dependent de ce rulează pe mașina care execută
testele.

Contextul _loop_isolated_from_the_network lasă un singur punct de contact fals,
send_heartbeat, și înlocuiește restul ușilor către rețea cu funcții care ridică
AssertionError. Sigilarea se face la frunzele de transport, nu la
register_agent_with_retry, ca orice cale viitoare din buclă să fie prinsă, iar
mesajul de eroare indică soluția pentru un test care chiar vrea calea de
reînregistrare.

+++ agent.py
@@ -327,7 +327,7 @@ def heartbeat_loop(
     config: Dict[str, Any],
     server_url: str,
     system_info: Dict[str, Any],
-    heartbeat_interval_seconds: int,
+    heartbeat_interval_seconds: float,
     stop_event: StopSignal,
 ) -> None:
     """
